Update data files
The way I compile data is I try to find airports for every IATA code.  If I can't find one, then I search the local codes as well.  (For example, STT is under the local code section).  What that means is that some parts of the data could be wrong, but if they are it would be tiny airports that people haven't heard of before so it should be relatively okay.